footballdb_scraper/requirements.txt
requests>=2.31.0
beautifulsoup4>=4.12.0
pandas>=2.0.0
pyyaml>=6.0
pyarrow>=14.0.0
urllib3>=2.0.0
lxml>=4.9.0

